uhidd should check if an HID boot device (keyboard or mouse) is in boot protocol mode. If so, uhidd should try to switch it to report protocol mode.
When an HID boot device is in boot protocol mode, it might only send
out standard(basic) reports that're understood by BIOS. As a result,
extra keys/buttons are most likely disabled. uhidd should set the
device in report protocol mode if it wants to support those extra
keys/buttons.

This change fixed the support for the multimedia keys found on
Microsoft Natural Ergonomic Keyboard 4000. A quirk added in FreeBSD
9.1 for the ukbd(4) kernel driver sets this keyboard to boot protocol
mode and multimedia keys (the second interface) are disabled as a
result.